update release flow
