--- scripts/test.main.kts
@@ -1,6 +1,0 @@
-#!/usr/bin/env kotlin
-
-@file:DependsOn("org.jetbrains.kotlin:kotlin-stdlib:1.7.0")
-
-println(KotlinVersion.CURRENT.toString())
-
